requirements.txt
# MCP Server Dependencies
mcp>=0.9.0
websockets>=12.0
pydantic>=2.0.0

# Database
SQLAlchemy>=2.0.0

# Authentication
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0
